+++ src/RadixDlt.NetworkGateway.PostgresIntegration/CommonDbContext.cs
@@ -147,6 +147,8 @@ internal abstract class CommonDbContext : DbContext
 
     public DbSet<KeyValueStoreEntryHistory> KeyValueStoreEntryHistory => Set<KeyValueStoreEntryHistory>();
 
+    public DbSet<KeyValueStoreAggregateHistory> KeyValueStoreAggregateHistory => Set<KeyValueStoreAggregateHistory>();
+
     public DbSet<ValidatorEmissionStatistics> ValidatorEmissionStatistics => Set<ValidatorEmissionStatistics>();
 
     public DbSet<NonFungibleSchemaHistory> NonFungibleSchemaHistory => Set<NonFungibleSchemaHistory>();
@@ -516,6 +518,10 @@ internal abstract class CommonDbContext : DbContext
         modelBuilder
             .Entity<KeyValueStoreEntryHistory>()
             .HasIndex(e => new { e.KeyValueStoreEntityId, e.Key, e.FromStateVersion });
+
+        modelBuilder
+            .Entity<KeyValueStoreAggregateHistory>()
+            .HasIndex(e => new { e.KeyValueStoreEntityId, e.FromStateVersion });
 
         modelBuilder
             .Entity<KeyValueStoreSchemaHistory>()