+++ src/RadixDlt.NetworkGateway.PostgresIntegration/LedgerExtension/SequencesHolder.cs
@@ -127,4 +127,6 @@ internal class SequencesHolder
     public long PackageBlueprintAggregateHistorySequence { get; set; }
 
     public long PackageCodeAggregateHistorySequence { get; set; }
+
+    public long KeyValueStoreAggregateHistorySequence { get; set; }
 }