+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Release Date: _unreleased_
   - added `manifest_instructions` optional property and a corresponding opt-in for returning original manifest of user transactions,
   - added optional `manifest_classes` property: a collection of zero or more manifest classes ordered from the most specific class to the least specific one.
 - added `permanently_rejects_at_epoch` to `/transaction/status` response for pending transactions. 
+- Added new endpoint `/state/key-value-store/keys/` that allows iterating over `KeyValueStore` keys.
 
 ### Deprecations
 - Obsoleted incorrectly named `access_rules_package` in favor of `role_assignment_module_package` on `NetworkConfigurationResponse.well_known_addresses`. Obsoleted property will contain effective copy of the new one for backwards compability.